pyproject.toml
[project]
name = "flydesk"
version = "0.1.0"
description = "Firefly Desk -- Backoffice as Agent"
readme = "README.md"
license = "Apache-2.0"
requires-python = ">=3.13"
authors = [
    { name = "Firefly Software Solutions Inc" },
]

dependencies = [
    "fireflyframework-genai[rest,security,postgres,http]",
    "authlib>=1.3.0",
    "httpx>=0.27.0",
    "asyncpg>=0.30.0",
    "aiosqlite>=0.22.0",
    "sqlalchemy[asyncio]>=2.0.0",
    "alembic>=1.14.0",
    "pgvector>=0.3.0",
    "redis>=5.0.0",
    "pydantic-settings>=2.0.0",
    "uvicorn>=0.30.0",
    "PyJWT[crypto]>=2.8",
    "html2text>=2024.2",
]

[project.scripts]
flydesk-seed = "flydesk.seeds.cli:main"
flydesk-serve = "flydesk.cli:serve"

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.24",
    "pytest-cov>=5.0",
    "ruff>=0.8",
    "pyright>=1.1",
    "testcontainers[postgres,redis]>=4.0",
    "httpx",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/flydesk"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[tool.ruff]
target-version = "py313"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP", "B", "A", "SIM", "TCH"]

[tool.uv]
python-preference = "only-system"

[tool.uv.sources]
fireflyframework-genai = { path = "../fireflyframework-genai", editable = true }

[tool.pyright]
pythonVersion = "3.13"
typeCheckingMode = "basic"

[dependency-groups]
dev = [
    "aiosqlite>=0.22.1",
]
